Feature: Medical Review PNC Mother Summary UI Designing 	- Privacy policy Implementations 	- Network check integrated 	- Session expire navigation to login activity 	- MDT-15623, MDT-15574

+++ Spice-SL/app/src/main/java/com/medtroniclabs/spice/db/entity/PregnancyDetail.kt
@@ -17,7 +17,7 @@ data class PregnancyDetail(
     // Pnc Detail
     var pncVisitNo: Long? = null,
     var dateOfDelivery: String? = null,
-    var noOfNeonates: Long? = null,
+    var noOfNeonates: Int? = null,
     var neonatePatientId: String? = null,
     var isDeliveryAtHome: Boolean? = null,
     //Childhood Visit Detail
